Tried Improving Input Path Validation
I don't know if u needed this but just felt like giving this a try

I made a total of 4 changes. First, added the "import sys" statement in the beginning. 
Second, a new code block in the main function to explicitly check the existence of the subject image file and the base model dir.
 
If any path is invalid, an error will be printed to the console, and the script is terminated preventing runtime errors later in the model loading process.